Add task to collate full tract tables for QA
The purpose of these tasks is, for a given tract, to read in all of the per-patch
{coaddName}Coadd_obj tables, select on a subset of columns that are useful for
QA analyses and plotting routines, aggregate these by tract and persist them as
datasets.  A separate table is created for each band.  These serve the purpose
of providing columns that are useful for QA but are not (yet?) included in the
DPDD-based per-tract objectTable_tract tables.  For ease of use with along with
the objectTable_tract tables (e.g. being able to use a single boolean expression
to filter on both), the columns of these new tables are sorted in the same order
as in the objectTable_tract tables and, since we don't want to make assumptions
on that ordering, we load in the table itself to do the sorting.

There are two tasks, makeForcedQaTractTables and makeUnforcedQaTractTables,
defined, the former to collect measurement values from the "forced_src" tables
(while including various columns from the "meas" and "ref" catalogs that are
useful for QA but do not get propagated to the "forced_src" catalogs), the
latter to collect measurement values from the "meas", i.e. unforced tables.