Fix band list for CoaddPlotFlag & Sn Selectors
The default for the bands list is already ["i"] for the SnSelector,
so it should not be included here.  On the other hand, the default
for the CoaddPlotFlagSelector is ["g", "r", "i", "z", "y"] and the
plots will end up empty if not all of these bands are present (even
though only the i band is being plotted), so these do need to be
overridden here.  This was noticed when running ci_hsc_gen3 which
only includes r and i band data, so plots for this pipeline were all
empty.