Use intended column name for CircAp12 plots
Any intended use for the "parameters" definitions are not being
taken advantage of and led to an incorrect column to be used for
one set of sky plots.  The safest option for now is to get rid
of the "parameters" in favor of explicitly listing the column
name (this way it should be easier to spot the incorrect use
of "i_ap09Flux" for a plot explicitly labeled as plot_CircAp12_*
that is being fixed here).