Add option to make colorColorPlots contour plots
Adds config parameter, contourPlot to ColorColorPlotTask such that,
when True, instead of a scatter plot color-mapped by magnitude, a
number density based contour plot is created with two panels
(left: stars, right: galaxies).

These take some time to make so should only be generated for specific
use cases or if runtime is not an issue (i.e. they are likely tier 2
or 3 plots).