Add defaults to classes and clean up pipelines
Many config settings were identical for all visit/coadd plots and
were being set for each plot definition in the associated pipelines.
Setting the defaults in the tasks allows for much simplified plot
definitions in the pipelines.  This also makes use of some
interdependencies to avoid configs getting out of sync (in particular
for the axis labels which are based on set parameters).