Add plotFlagSelector in addition to flagSelector
This renames the previous "flagSelector" which defaulted to
Coadd[Visit]PlotFlagSelector to "plotFlagSelector" and adds a "flagSelector"
that defaults to the FlagSelector.  This way we can add flags that are not
included in the fixed set in Coadd[Visit]PlotFlagSelector.  In cases where
the current defaults are not appropriate, a given pipeline definition should
override the selectorActions.flagSelector.selectWhenFalse/True fields.  If
no extra flagging is desired, this can be accomplished by setting the above
fields to an empty list, or by putting, e.g.

selectorActions.remove: ["nameOfSelector", "nameOfAnotherSelector"]

in the config section of the pipeline.

The need to add this was noticed when the stellar locus fit plot was failing
with NaNs when using CModel flux (only psfFlux gets flagged in the
Coadd[Visit]PlotFlagSelector, so CModel failures could sneak in).  The issue
is actually widespread and remedied here.