Divide coadd pipelines into tiers
In particular, remove all tier 2 and tier 3 plots from coaddQAPlot.yaml
and coaddQAEllipPlot.yaml and move them into a new coaddQAPlots_tier2_3.yaml
that is not meant to be run in regular production.  In order to access them,
a pipelines/HSC/analysis_drp_plots_tier2_3.yaml is also being added to the
drp_pipe repo.