Updates and fixes for stellar locus fit plotting
This includes setting configs more consistently, fixing issues where
the plot was misleading (i.e. not reflecting reality in terms of the
points being used in the fit and the density color-mapping), updating
the statistics to be given in mmag, updating labels to add info as
well as remove some irrelevant entries, and adds the ability to set
axis limits explicitly.